feat: added new parameter rep ASSIGNMENT_FORM to leave params in brackets with text (#1193)
* initial refactor to controlinterface


* changed parameter format to ASSIGNMENT_FORM, refactored ControlInterface, made some private functions public
