+++ trestle/core/commands/author/jinja.py
@@ -33,7 +33,7 @@ from trestle.common.model_utils import ModelUtils
 from trestle.core.catalog_interface import CatalogInterface
 from trestle.core.commands.command_docs import CommandPlusDocs
 from trestle.core.commands.common.return_codes import CmdReturnCodes
-from trestle.core.control_interface import ControlInterface
+from trestle.core.control_interface import ControlInterface, ParameterRep
 from trestle.core.docs_control_writer import DocsControlWriter
 from trestle.core.jinja import MDCleanInclude, MDDatestamp, MDSectionInclude
 from trestle.core.profile_resolver import ProfileResolver
@@ -177,7 +177,7 @@ class JinjaCmd(CommandPlusDocs):
             profile_path = ModelUtils.full_path_for_top_level_model(trestle_root, profile, Profile)
             profile_resolver = ProfileResolver()
             resolved_catalog = profile_resolver.get_resolved_profile_catalog(
-                trestle_root, profile_path, False, False, parameters_formatting
+                trestle_root, profile_path, False, False, parameters_formatting, ParameterRep.ASSIGNMENT_FORM
             )
 
             ssp_writer = SSPMarkdownWriter(trestle_root)
@@ -215,7 +215,7 @@ class JinjaCmd(CommandPlusDocs):
         profile, profile_path = ModelUtils.load_top_level_model(trestle_root, profile_name, Profile)
         profile_resolver = ProfileResolver()
         resolved_catalog = profile_resolver.get_resolved_profile_catalog(
-            trestle_root, profile_path, False, False, parameters_formatting
+            trestle_root, profile_path, False, False, parameters_formatting, ParameterRep.ASSIGNMENT_FORM
         )
         catalog_interface = CatalogInterface(resolved_catalog)
